Refactore Models Classes: Robot and Drone
