+++ src/decentralised_adaptive_coverage/scripts/config.ini
@@ -9,18 +9,16 @@ num_particles = 1000
 
 [VORONOI_SETUP]
 boundary_point_resolution = 100
+boundary_tolerance = 0.02
 
 [ITERATIVE_PROCESS]
-time_step = 0.05
-time_limit = 5
-beta = 0.1
-t_final = 10
-n_iterations = 1
+n_iterations = 2
 distance_threshold = 1.0
 
 [FILTER_SETUP]
 filter_type = UKF
 sampling_time = 30  
+apply_consensus = True
 
 [SIMULATION_SETUP]
 sampling_time = 10 
@@ -30,7 +28,7 @@ time_per_step = 1
 parallelism_mode = threading
 
 [GAZEBO_SETUP]
-enable_physics_simulation = True
+enable_physics_simulation = False
 world_name = farm
 home_area = 9
